refactor(package): move adapters into view/adapter
