+++ app/src/main/java/com/example/ezvault/view/adapter/ItemAdapter.java
@@ -1,4 +1,4 @@
-package com.example.ezvault;
+package com.example.ezvault.view.adapter;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -15,6 +15,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.ezvault.R;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.utils.ItemListView;
